+++ packages/core/src/index.ts
@@ -7,3 +7,4 @@ export * from './lib/form-persistence.service';
 export * from './lib/providers';
 export * from './lib/presets';
 export * from './lib/define-form';
+export * from './lib/json-schema';